+++ CHANGELOG.md
@@ -3,6 +3,7 @@ Changelog
 
 v2.3.0
 ------
+* Fixed a bug when a restored object has access to a method which wasn't originally implemented.
 * Replaced the Session.stateListener method with addStateListener.
 * Added a debug option to close pipes when they are recycled.
 * Methods unimplemented on the server throw an exception only allowed by the RemoteFailure